admin role

+++ src/main/java/com/wawrze/library/services/BookDbService.java
@@ -1,6 +1,7 @@
 package com.wawrze.library.services;
 
 import com.wawrze.library.domains.books.Book;
+import com.wawrze.library.exeptions.ForbiddenException;
 import com.wawrze.library.repositories.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,11 +29,13 @@ public class BookDbService {
         return bookRepository.findById(id);
     }
 
-    public Book saveBook(final Book book) {
+    public Book saveBook(final Book book, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage books!");
         return bookRepository.save(book);
     }
 
-    public void deleteBook(final Integer id) {
+    public void deleteBook(final Integer id, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage books!");
         bookRepository.delete(id);
     }
