+++ src/main/java/com/wawrze/library/services/RentDbService.java
@@ -1,6 +1,7 @@
 package com.wawrze.library.services;
 
 import com.wawrze.library.domains.rents.Rent;
+import com.wawrze.library.exeptions.ForbiddenException;
 import com.wawrze.library.repositories.RentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,19 +21,23 @@ public class RentDbService {
         this.rentRepository = rentRepository;
     }
 
-    public List<Rent> getAllRents() {
+    public List<Rent> getAllRents(final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage rents!");
         return rentRepository.findAll();
     }
 
-    public Optional<Rent> getRent(final Integer id) {
+    public Optional<Rent> getRent(final Integer id, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage rents!");
         return rentRepository.findById(id);
     }
 
-    public Rent saveRent(final Rent rent) {
+    public Rent saveRent(final Rent rent, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage rents!");
         return rentRepository.save(rent);
     }
 
-    public void deleteRent(final Integer id) {
+    public void deleteRent(final Integer id, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage rents!");
         rentRepository.delete(id);
     }
