+++ src/main/java/com/wawrze/library/services/UserDbService.java
@@ -1,7 +1,8 @@
 package com.wawrze.library.services;
 
-import com.wawrze.library.domains.users.LoginDto;
+import com.wawrze.library.domains.users.Credentials;
 import com.wawrze.library.domains.users.User;
+import com.wawrze.library.domains.users.UserRole;
 import com.wawrze.library.exeptions.ForbiddenException;
 import com.wawrze.library.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,8 +14,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import static com.wawrze.library.filters.AuthFilter.TOKEN_KEY;
-import static com.wawrze.library.filters.AuthFilter.USER_ID_KEY;
+import static com.wawrze.library.filters.AuthFilter.*;
 
 @Service
 @Transactional
@@ -25,35 +25,56 @@ public class UserDbService {
     @Autowired
     public UserDbService(UserRepository userRepository) {
         this.userRepository = userRepository;
+        createAdminIfNeeded();
     }
 
-    public List<User> getAllUsers() {
+    public List<User> getAllUsers(final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage users!");
         return userRepository.findAll();
     }
 
-    public Optional<User> getUser(final Integer id) {
+    public Optional<User> getUser(final Integer id, final int userId, final boolean isAdmin) {
+        if (id != userId && !isAdmin) throw new ForbiddenException("Only administrator can manage users!");
         return userRepository.findById(id);
     }
 
-    public User saveUser(final User user) {
+    public User saveUser(final User user, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage users!");
         return userRepository.save(user);
     }
 
-    public void deleteUser(final Integer id) {
+    public void deleteUser(final Integer id, final boolean isAdmin) {
+        if (!isAdmin) throw new ForbiddenException("Only administrator can manage users!");
         userRepository.delete(id);
     }
 
-    public String login(final LoginDto loginDto, HttpServletRequest request) {
-        User user = userRepository.findByLogin(loginDto.getLogin());
+    public String login(final Credentials credentials, HttpServletRequest request) {
+        User user = userRepository.findByLogin(credentials.getLogin());
 
         if (user == null) throw new ForbiddenException("No user!");
-        if (!user.getPassword().equals(loginDto.getPassword())) throw new ForbiddenException("Wrong password!");
+        if (!user.getPassword().equals(credentials.getPassword())) throw new ForbiddenException("Wrong password!");
 
         String token = UUID.randomUUID().toString();
         request.getSession().setAttribute(USER_ID_KEY, user.getId());
         request.getSession().setAttribute(TOKEN_KEY, token);
+        request.getSession().setAttribute(IS_ADMIN_KEY, user.getUserRole() == UserRole.ADMIN);
 
         return token;
+    }
+
+    private void createAdminIfNeeded() {
+        if (userRepository.findAll().size() == 0) {
+            User admin = new User(
+                    null,
+                    "admin",
+                    "admin",
+                    "Admin",
+                    "Admin",
+                    null,
+                    UserRole.ADMIN
+            );
+            userRepository.save(admin);
+        }
     }
 
 }
